% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNroyDesignPoints.R
\name{addNroyDesignPoints}
\alias{addNroyDesignPoints}
\title{Augment a design with NROY design points}
\usage{
addNroyDesignPoints(
  X,
  Y,
  Y_target,
  n_aug,
  mins_aug,
  maxes_aug,
  thres = 3,
  disc_list,
  disc_sd_list,
  obs_sd_list,
  n_app,
  nreps,
  shrink = TRUE
)
}
\arguments{
\item{X}{original design matrix}

\item{Y}{model output matrix}

\item{Y_target}{target output or observation}

\item{n_aug}{number of candidate points with which to augment the design}

\item{mins_aug}{minima of n_aug}

\item{maxes_aug}{maxima of n_aug}

\item{thres}{implausibility threshold}

\item{disc_list}{model discrepancy (bias)}

\item{disc_sd_list}{model discrepancy uncertainty}

\item{obs_sd_list}{observation (Y_target) uncertainty}

\item{n_app}{final number of points with which to augment the design}

\item{nreps}{number of iterations in the maximin search}

\item{shrink}{keep only members of the original design hat fall within marginal limits of the NROY}
}
\value{
a list
}
\description{
Augment a design with NROY design points
}
