% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_bt.R
\name{all_bt}
\alias{all_bt}
\title{Are all elements of a vector below a threshold?}
\usage{
all_bt(x, thres)
}
\arguments{
\item{x}{vector}

\item{thres}{threshold}
}
\value{
logical
}
\description{
Are all elements of a vector below a threshold?
}
\examples{
all_bt(c(1,2,3), 3)
}
