% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_shadow.R
\name{bind_shadow}
\alias{bind_shadow}
\title{Bind a shadow matrix to a data frame with NA values.}
\usage{
bind_shadow(data, vars = dplyr::everything())
}
\arguments{
\item{data}{A data frame with missing values.}

\item{vars}{A subset of variables to include in the shadow matrix.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
The bind_shadow function binds a shadow matrix to the data frame.
}
