% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseMaximinNroy.R
\name{chooseMaximinNroy}
\alias{chooseMaximinNroy}
\title{Choose a set of NROY points with the largest minimum distrance}
\usage{
chooseMaximinNroy(n_app, waveobj, nreps)
}
\arguments{
\item{n_app}{desired size of new design}

\item{waveobj}{wave object}

\item{nreps}{repetitions}
}
\value{
matrix of design points
}
\description{
Choose a set of NROY points with the largest minimum distrance
}
