% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createKmFitListParallel.R
\name{createKmFitListParallel}
\alias{createKmFitListParallel}
\title{Create a list of dicekriging fits using parallel processing}
\usage{
createKmFitListParallel(X, Y, formula = ~., multistart = 4, ...)
}
\arguments{
\item{X}{X matrix of inputs (design), variables in columns}

\item{Y}{Y matrix of outputs, variables in columns}

\item{formula}{formula for the prior km fit, with a default as a linear model '~.'}

\item{multistart}{number of starts for the hyperparameter optimisation}

\item{...}{}
}
\value{
list of km (dicekriging) model fits, with each element of the list corresponding to a column of Y.
}
\description{
Create a list of dicekriging fits using parallel processing
}
