% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shadow.R
\name{get_shadow}
\alias{get_shadow}
\title{Create a shadow matrix based on the missing values in a data frame.}
\usage{
get_shadow(data, vars = dplyr::everything())
}
\arguments{
\item{data}{A data frame with missing values.}

\item{vars}{A subset of variables to include in the shadow matrix.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
The get_shadow function builds a shadow matrix based on the
passed in data frame.
}
