% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEmulatorPostVar.R
\name{getEmulatorPostVar}
\alias{getEmulatorPostVar}
\title{Returns the min, mean and max emulator posterior variance for each output in a wave object}
\usage{
getEmulatorPostVar(waveobj)
}
\arguments{
\item{waveobj}{}
}
\value{

}
\description{
Returns the min, mean and max emulator posterior variance for each output in a wave object
}
