% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl.R
\name{impl}
\alias{impl}
\title{Calculate Implausibility}
\usage{
impl(em, em.sd, disc, disc.sd, obs, obs.sd)
}
\arguments{
\item{em}{Emulator mean}

\item{em.sd}{Emulator uncertainty (standard deviation)}

\item{disc}{Model discrepancy mean}

\item{disc.sd}{Model discrepancy uncertainty (standard deviation)}

\item{obs}{Observation mean}

\item{obs.sd}{Observation uncertainty (standard deviation)}
}
\value{
Implausibility score
}
\description{
Calculate Implausibility
}
