% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2list.R
\name{mat2list}
\alias{mat2list}
\title{Create a list of each column of a matrix}
\usage{
mat2list(X)
}
\arguments{
\item{X}{}
}
\value{
list of each column of X
}
\description{
Create a list of each column of a matrix
}
