% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot density of observed and imputed values for a single variable}
\usage{
plot_density(data, imp_id, plot_var, kernel = "auto")
}
\arguments{
\item{data}{A data frame containing multiple imputations.}

\item{imp_id}{A grouping column that identifies the rows belonging to each
imputation.}

\item{plot_var}{The variable you wish to plot.}

\item{kernel}{The kernel density type you want to use. "auto" will select
Rectangular for variables with < 5 unique values and Gaussian otherwise.}
}
\description{
This function returns density plots for the observed data (blue) and plots for
the imputed values of each imputation (red).
}
