% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_r.R
\name{pool.r}
\alias{pool.r}
\title{Pool Pearson correlation coefficients}
\usage{
pool.r(rVec = NULL, N = NULL)
}
\arguments{
\item{rVec}{A vector of correlation coefficients}

\item{N}{The number of imputations}
}
\value{
A vector of pooled r's
}
\description{
Pool Pearson correlation coefficients
}
