% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_thres.R
\name{prop_thres}
\alias{prop_thres}
\title{Proportion of elements of a vector below a threshold}
\usage{
prop_thres(x, thres, above = FALSE)
}
\arguments{
\item{x}{vector}

\item{thres}{threshold}

\item{above}{logical: should the function calculate the proportion above the threshold?}
}
\value{

}
\description{
Proportion of elements of a vector below a threshold
}
\examples{
prop_thres(x = 1:10, thres = 3, above = FALSE)
}
