% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_mids.R
\name{stack_mids}
\alias{stack_mids}
\title{Convert a list of data.frames to a stacked tibble}
\usage{
stack_mids(mids, .setid = "stack_id")
}
\arguments{
\item{mids}{A mice mids object.}

\item{.setid}{Set the name of the key column. In cases where there are
more than 1 data.frames the key column increments by 1 for each data set.
Defaults to "stack_id"}
}
\description{
Convert a list of data.frames to a stacked tibble
}
\examples{
stack <- stack_mids(mids, .setid = "id_column")
stack <- stack_mids(mids)
}
