% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withinRange.R
\name{withinRange}
\alias{withinRange}
\title{Elements of a vector that fall between thresholds}
\usage{
withinRange(x, maxes, mins)
}
\arguments{
\item{x}{vector}

\item{maxes}{maxima}

\item{mins}{minima}
}
\value{
logical vector of length one.
}
\description{
Elements of a vector that fall between thresholds
}
