% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl.R
\name{impl}
\alias{impl}
\title{Calculate Implausibility}
\usage{
impl(em, em_sd, disc, disc_sd, obs, obs_sd)
}
\arguments{
\item{em}{Emulator mean}

\item{em_sd}{Emulator uncertainty (standard deviation)}

\item{disc}{Model discrepancy mean}

\item{disc_sd}{Model discrepancy uncertainty (standard deviation)}

\item{obs}{Observation mean}

\item{obs_sd}{Observation uncertainty (standard deviation)}
}
\value{
Implausibility score
}
\description{
Calculate Implausibility
}
